chore(tms): move shipment models to domain package
