+++ app/tms/shipment/models/__init__.py
@@ -1,0 +1,12 @@
+# app/tms/shipment/models/__init__.py
+# Domain-owned ORM models for TMS shipment and shipment preparation.
+
+from app.tms.shipment.models.order_shipment_prepare import OrderShipmentPrepare
+from app.tms.shipment.models.order_shipment_prepare_package import OrderShipmentPreparePackage
+from app.tms.shipment.models.transport_shipment import TransportShipment
+
+__all__ = [
+    "OrderShipmentPrepare",
+    "OrderShipmentPreparePackage",
+    "TransportShipment",
+]
