+++ app/tms/shipment/service_prepare_quotes.py
@@ -17,8 +17,8 @@ from fastapi import HTTPException, status
 from sqlalchemy import select, text
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.order_shipment_prepare import OrderShipmentPrepare
-from app.models.order_shipment_prepare_package import OrderShipmentPreparePackage
+from app.tms.shipment.models.order_shipment_prepare import OrderShipmentPrepare
+from app.tms.shipment.models.order_shipment_prepare_package import OrderShipmentPreparePackage
 from app.tms.quote.recommend import recommend_quotes
 from app.tms.quote.types import Dest
 from app.tms.quote_snapshot import build_quote_snapshot
